fluent_ui_ll

+++ nice_ui/ui/work_srt.py
@@ -3,21 +3,24 @@ import re
 import sys
 
 import path
-from PySide6.QtCore import Qt, QRect, Slot, QSettings
+from PySide6.QtCore import Qt, Slot, QSettings, QSize
 from PySide6.QtGui import QDragEnterEvent, QDropEvent
-from PySide6.QtWidgets import (QWidget, QVBoxLayout, QHBoxLayout, QFileDialog, QLabel, QTableWidget, QApplication, )
-from qfluentwidgets import PushButton, ComboBox, TableWidget, FluentIcon, MessageBox
+from PySide6.QtWidgets import (QWidget, QVBoxLayout, QHBoxLayout, QFileDialog, QTableWidget, QApplication, QAbstractItemView, QTableWidgetItem, QSizePolicy,
+                               QFormLayout, )
 
+from agent import get_translate_code
 from nice_ui.configure import config
 from nice_ui.main_win.secwin import SecWindow
-from videotrans.translator import TRANSNAMES
+from orm.queries import PromptsOrm
+from vendor.qfluentwidgets import PushButton, TableWidget, FluentIcon, InfoBar, InfoBarPosition, ComboBox, BodyLabel
 
 
 class WorkSrt(QWidget):
-    def __init__(self, text: str, parent=None, setting=None):
+    def __init__(self, text: str, parent=None, settings=None):
         super().__init__(parent=parent)
-        self.setting = setting
-        self.table = LTableWindow(self, setting)
+        self.settings = settings
+        self.prompts_orm = PromptsOrm()
+        self.table = LTableWindow(self, settings)
         self.util = SecWindow(self)
         self.language_name = config.langnamelist
         self.setObjectName(text.replace(" ", "-"))
@@ -25,8 +28,6 @@ class WorkSrt(QWidget):
         self.bind_action()
 
     def setup_ui(self):
-        self.setWindowTitle("字幕翻译准备")
-        self.setGeometry(QRect(12, 16, 777, 578))
 
         main_layout = QVBoxLayout()
         self.setLayout(main_layout)
@@ -37,8 +38,9 @@ class WorkSrt(QWidget):
         self.btn_get_srt.setFixedHeight(100)  # 增加按钮的高度
         main_layout.addWidget(self.btn_get_srt)
 
-        self.add_queue_btn = PushButton("添加到队列")
-        main_layout.addWidget(self.add_queue_btn)
+        # # todo: 增加到队列按钮到时候屏蔽掉
+        # self.add_queue_btn = PushButton("添加到队列")
+        # main_layout.addWidget(self.add_queue_btn)
 
         # 下拉框布局
         combo_layout = QHBoxLayout()
@@ -48,7 +50,7 @@ class WorkSrt(QWidget):
         source_layout = QHBoxLayout()
         source_layout.setAlignment(Qt.AlignmentFlag.AlignLeading | Qt.AlignmentFlag.AlignLeft | Qt.AlignmentFlag.AlignVCenter)
 
-        source_language_name = QLabel("原始语种")
+        source_language_name = BodyLabel("原始语种")
 
         self.source_language_combo = ComboBox(self)
         self.source_language_combo.addItems(self.language_name)
@@ -64,7 +66,7 @@ class WorkSrt(QWidget):
         translate_layout = QHBoxLayout()
         translate_layout.setAlignment(Qt.AlignmentFlag.AlignLeading | Qt.AlignmentFlag.AlignLeft | Qt.AlignmentFlag.AlignVCenter)
 
-        translate_language_name = QLabel("翻译语种")
+        translate_language_name = BodyLabel("翻译语种")
 
         self.translate_language_combo = ComboBox(self)
         self.translate_language_combo.addItems(self.language_name)
@@ -79,64 +81,102 @@ class WorkSrt(QWidget):
         engine_layout = QHBoxLayout()
         engine_layout.setAlignment(Qt.AlignmentFlag.AlignLeading | Qt.AlignmentFlag.AlignLeft | Qt.AlignmentFlag.AlignVCenter)
 
-        translate_model_name = QLabel("翻译引擎")
+        translate_model_name = BodyLabel("翻译引擎")
 
         self.translate_model = ComboBox(self)
-        self.translate_model.addItems(TRANSNAMES)
-        translate_name = (config.params["translate_type"] if config.params["translate_type"] in TRANSNAMES else TRANSNAMES[0])
-
+        # todo: 翻译引擎列表需调整
+        translate_list = get_translate_code()
+        self.translate_model.addItems(translate_list)
+        translate_name = config.params["translate_type"]
+        config.logger.info(f"translate_name: {translate_name}")
         self.translate_model.setCurrentText(translate_name)
 
         engine_layout.addWidget(translate_model_name)
         engine_layout.addWidget(self.translate_model)
         combo_layout.addLayout(engine_layout)
 
+        # todo: 只有选择ai时才显示
+        prompt_layout = QHBoxLayout()
+        prompt_layout.setAlignment(Qt.AlignmentFlag.AlignLeading | Qt.AlignmentFlag.AlignLeft | Qt.AlignmentFlag.AlignVCenter)
+        ai_prompt_name = BodyLabel("提示词")
+        self.ai_prompt = ComboBox(self)
+        self.ai_prompt.addItems(self._get_ai_prompt())
+        self.ai_prompt.setCurrentText(config.params["prompt_name"])
+        prompt_layout.addWidget(ai_prompt_name)
+        prompt_layout.addWidget(self.ai_prompt)
+        main_layout.addLayout(prompt_layout)
+
         # 表格
+        media_table_layout = QHBoxLayout()
+        media_table_layout.setContentsMargins(60, -1, 60, -1)
         self.media_table = TableWidget(self)
         self.media_table.setColumnCount(5)
         self.media_table.setHorizontalHeaderLabels(['文件名', '字符数', '算力消耗', '操作', '文件路径'])
-        self.media_table.setColumnWidth(0, 50)
-        self.media_table.setColumnWidth(1, 200)
-        self.media_table.setColumnWidth(2, 200)
+        self.media_table.setColumnWidth(0, 400)
+        self.media_table.setColumnWidth(1, 100)
+        self.media_table.setColumnWidth(2, 100)
         self.media_table.setColumnWidth(3, 100)
         self.media_table.setColumnWidth(4, 100)
-        self.media_table.setShowGrid(False)
-        main_layout.addWidget(self.media_table)
+        # self.media_table.setShowGrid(False) #隐藏网格线
+        self.media_table.setColumnHidden(2, True)  # 隐藏操作列
+        self.media_table.setColumnHidden(4, True)  # 隐藏操作列
+        self.media_table.setEditTriggers(QAbstractItemView.NoEditTriggers)
+        media_table_layout.addWidget(self.media_table)
+        main_layout.addLayout(media_table_layout)
 
         # 开始按钮
-        self.start_button = PushButton("开始翻译", self)
+        self.formLayout_5 = QFormLayout()
+        self.formLayout_5.setFormAlignment(Qt.AlignmentFlag.AlignCenter)
+        self.formLayout_5.setContentsMargins(-1, -1, -1, 20)
+        self.start_button = PushButton("开始", self)
         self.start_button.setIcon(FluentIcon.PLAY)
 
-        main_layout.addWidget(self.start_button)
+        sizePolicy5 = QSizePolicy(QSizePolicy.Policy.Fixed, QSizePolicy.Policy.Fixed)
+        sizePolicy5.setHorizontalStretch(0)
+        sizePolicy5.setVerticalStretch(0)
+        sizePolicy5.setHeightForWidth(self.start_button.sizePolicy().hasHeightForWidth())
+        self.start_button.setSizePolicy(sizePolicy5)
+        self.start_button.setMinimumSize(QSize(200, 50))
+        self.formLayout_5.setWidget(0, QFormLayout.LabelRole, self.start_button)
+
+        main_layout.addLayout(self.formLayout_5)
 
         # # 设置接受拖放  # self.setAcceptDrops(True) #不知道为啥不好使了
 
     def bind_action(self):
-        self.start_button.clicked.connect(self.start_translation)
+
         # 选择文件,并显示路径
         self.btn_get_srt.clicked.connect(self.table.select_file)
         self.btn_get_srt.setAcceptDrops(True)
         self.btn_get_srt.dragEnterEvent = self.table.drag_enter_event
         self.btn_get_srt.dropEvent = lambda event: self.table.drop_event(self.media_table, event)
-
-        self.add_queue_btn.clicked.connect(self.add_queue_srt)
+        # self.add_queue_btn.clicked.connect(self.add_queue_srt)
+        self.start_button.clicked.connect(self.util.check_translate)
 
     def add_queue_srt(self):
         # 获取self.main.media_table中第4列的数据
         srt_list = []
         for i in range(self.media_table.rowCount()):
-            srt_list.append(self.media_table.cellWidget(i, 4).text())
+            srt_list.append(self.media_table.item(i, 4).text())
         config.queue_srt.extend(srt_list)
         config.logger.info(f"queue_srt: {config.queue_srt}")
 
-    def start_translation(self):
-        MessageBox("提示", "翻译准备完成，即将开始翻译...", self).exec()
+
+
+    def _get_ai_prompt(self) -> list:
+        # 获取AI提示列表
+        prompt_names = self.prompts_orm.get_prompt_name()
+        prompt_name_list = []
+        for i in prompt_names:
+            prompt_name_list.append(i.prompt_name)
+        return prompt_name_list
+
 
 
 class LTableWindow:
-    def __init__(self, main, setting):
+    def __init__(self, main, settings):
         self.main = main
-        self.setting = setting
+        self.settings = settings
 
     def select_file(self):
         file_paths = QFileDialog()
@@ -149,63 +189,57 @@ class LTableWindow:
             for file_path in file_paths:
                 self.add_file_to_table(self.main.media_table, file_path)
         config.last_opendir = os.path.dirname(file_paths[0])
-        self.setting.setValue("last_dir", config.last_opendir)
+        self.settings.setValue("last_dir", config.last_opendir)
 
     # 列表的操作
-    def table_set_item(self, ui_table, row_position: int, l0, l1, l2, ):
-        #文件名
-        file_name = QLabel()
-        file_name.setText(os.path.basename(l0))
-        file_name.setAlignment(Qt.AlignCenter)
-        ui_table.setCellWidget(row_position, 0, file_name)
-
-        # 字符数
-        file_character_count = QLabel()
-        file_character_count.setText(l1)
-        file_character_count.setAlignment(Qt.AlignCenter)
-        ui_table.setCellWidget(row_position, 1, file_character_count)
-
-        # 算力消耗
-        locol_value = QLabel()
-        locol_value.setText(l2)
-        locol_value.setAlignment(Qt.AlignCenter)
-        ui_table.setCellWidget(row_position, 2, locol_value)
-
-        # 操作
-        delete_button = PushButton("删除")
-        delete_button.setStyleSheet("background-color: red; color: white;")
-        ui_table.setCellWidget(row_position, 3, delete_button)
-        delete_button.clicked.connect(lambda _, row=row_position: self.delete_file(ui_table, row))
-
-        # 文件路径
-        file_path = QLabel()
-        file_path.setText(l0)
-        file_path.setAlignment(Qt.AlignCenter)
-        ui_table.setCellWidget(row_position, 4, file_path)
 
     def add_file_to_table(self, ui_table: TableWidget, file_path: str):
         # 添加文件到表格
         row_position = ui_table.rowCount()
-        ui_table.insertRow(row_position)
-
-        file_name = os.path.basename(file_path)
-        config.logger.info(f"file_name: {file_name}")
         file_character_count = self.get_file_character_count(file_path)
-        config.logger.info(f"file_character_count: {file_character_count}")
-        config.logger.info(f"file_path: {file_path}")
+        if file_character_count:
+            ui_table.insertRow(row_position)
+            file_name = os.path.basename(file_path)
+            config.logger.info(f"file_name type: {type(file_name)}")
+            config.logger.info(f"file_name: {file_name}")
+            config.logger.info(f"file_character_count: {file_character_count}")
+            config.logger.info(f"file_path: {file_path}")
+            # todo: 算力消耗
+            # 文件名
+            ui_table.setItem(row_position, 0, QTableWidgetItem(file_name))
+            # 字符数
+            ui_table.setItem(row_position, 1, QTableWidgetItem(str(file_character_count)))
+            # 算力消耗
+            ui_table.setItem(row_position, 2, QTableWidgetItem("未知"))
+            # 操作
+            delete_button = PushButton("删除")
+            delete_button.setFixedSize(QSize(80, 30))
+            delete_button.setSizePolicy(QSizePolicy.Fixed, QSizePolicy.Fixed)  # 设置大小策略为Fixed
+
+            delete_button.setStyleSheet("background-color: #FF6C64; color: white;")
+            ui_table.setCellWidget(row_position, 3, delete_button)
+            delete_button.clicked.connect(lambda _, row=row_position: self.delete_file(ui_table, row))
+            # 文件路径
+            ui_table.setItem(row_position, 4, QTableWidgetItem(file_path))
+        else:
+            InfoBar.error(title='失败', content="文件内容错误，请检查文件内容", orient=Qt.Horizontal, isClosable=True, position=InfoBarPosition.TOP_RIGHT, duration=2000, parent=self.main)
 
     # 获取文件字符数
-    def get_file_character_count(self, file_path: path) -> str:
+    def get_file_character_count(self, file_path: path) -> int | bool:
         # 输入srt格式的文件，获取里面的字符数量，不计算序号，不计算时间戳
         character_count = 0
         with open(file_path, "r", encoding="utf-8") as f:
-            lines = f.readlines()
+            try:
+                lines = f.readlines()
+            except UnicodeDecodeError:
+                config.logger.error(f"文件{file_path}编码错误，请检查文件编码格式")
+                return False
             for line in lines:
                 # 跳过序号和时间戳
                 if re.match(r"^\d+$", line.strip()) or re.match(r"^\d{2}:\d{2}:\d{2},\d{3} --> \d{2}:\d{2}:\d{2},\d{3}$", line.strip(), ):
                     continue
                 character_count += len(line.strip())
-        return str(character_count)
+        return character_count
 
     @Slot()
     def delete_file(self, ui_table: QTableWidget, row: int):
